refactoring runtimes cmd to rack package & switching to registerwithoutprovider for cli cmd
