+++ pkg/cli/runtimes.go
@@ -1,21 +1,22 @@
 package cli
 
 import (
+	"github.com/convox/convox/pkg/rack"
 	"github.com/convox/convox/sdk"
 	"github.com/convox/stdcli"
 )
 
 func init() {
-	register("runtimes", "get list of runtimes", Runtimes, stdcli.CommandOptions{
+	registerWithoutProvider("runtimes", "get list of runtimes", Runtimes, stdcli.CommandOptions{
 		Usage:    "<orgname>",
 		Validate: stdcli.ArgsMin(1),
 	})
 }
 
-func Runtimes(runtime sdk.Interface, c *stdcli.Context) error {
+func Runtimes(_ sdk.Interface, c *stdcli.Context) error {
 	org := c.Arg(0)
 
-	rs, err := runtime.OrganizationRuntimes(org)
+	rs, err := rack.Listruntimes(c, org)
 	if err != nil {
 		return err
 	}
