+++ sdk/methods.go
@@ -868,18 +868,6 @@ func (c *Client) Runtimes(rackOrgSlug string) (structs.Runtimes, error) {
 	return v, err
 }
 
-func (c *Client) OrganizationRuntimes(org string) (structs.Runtimes, error) {
-	var err error
-
-	ro := stdsdk.RequestOptions{Headers: stdsdk.Headers{}, Params: stdsdk.Params{}, Query: stdsdk.Query{}}
-
-	var v structs.Runtimes
-	
-	err = c.Get(fmt.Sprintf("/organizations/%s/runtimes", org), ro, &v)
-
-	return v, err
-}
-
 func (c *Client) WorkflowList(rackOrgSlug string) (structs.WorkflowListResp, error) {
 	var err error
